Include possibility to filter feed entries by locales.
When using the multi-site feature of Statamic, there might be feed collections available in multiple languages. This feature adds the abilitiy to only show entries for a given set of locales.

The default behavior is to show entries of all locales (to ensure backwards compatibility).
To limit the entries locales, an array of locales can be provided.
To show entries of the current local only, the special keyword 'current' can be used.