Fix incorrect update date when building static site in CI
The feed currently uses the file's last modification time, which isn't a reliable indicator, especially when cloning the repo in a CI environment and will typically set the `updated` field for every entry in the field to the deploy date.

The `lastModified` method returns a Carbon instance created from the `updated_at` field on the entry if it exists. Otherwise, it falls back to using the `fileLastModified` method.